frontend/backend/requirements.txt
apple-music-python==1.0.5
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
filelock==3.15.4
fonttools==4.53.0
fuzzywuzzy==0.18.0
git-filter-repo==2.45.0
gunicorn==23.0.0
identify==2.5.36
idna==3.7
kiwisolver==1.4.5
Levenshtein==0.25.1
matplotlib==3.9.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
PyYAML==6.0.1
rapidfuzz==3.9.6
redis==5.0.5
requests==2.32.3
six==1.16.0
spotipy==2.24.0
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
virtualenv==20.26.3
